move vector to CSR.
